update multiple smaller changes, output bugs
git-svn-id: http://Marc-PC/svn/Full@1551 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/PnET-cohort-library/trunk/src/WrappedDisturbance.cs
@@ -46,9 +46,7 @@ namespace Landis.Library.BiomassCohortsPnET
         public int ReduceOrKillMarkedCohort(Landis.Library.BiomassCohorts.ICohort cohort)
         {
          
-            return (int)cohort.Biomass;
-            
-            //throw new System.Exception("Incompatibitlity");
+            return cohort.Biomass;
         }
         
         public int ReduceOrKillMarkedCohort(ICohort cohort)
